requirements.txt
python-docx
Pillow
pymupdf
